Added min size for app; factored hover description into its own hook.
